backend/services/image_explainer.py
from PIL import Image

class ImageExplainer:
    """
    Explainable AI service for medical imaging.
    Currently stubbed to support future integration of Grad-CAM
    or bounding-box algorithms overlaying diagnostic heatmaps.
    """

    async def generate_heatmap(self, base_image: Image.Image, text_context: str = "") -> dict:
        """
        Takes a raw patient image (e.g., X-ray) and outputs bounding regions
        or a heatmap overlay image based on detected anomalies.
        """
        
        # TODO: Integrate external Grad-CAM model or specific vision APIs
        # For now, returning dummy placeholder coordinates.
        
        return {
            "service": "ImageExplainer",
            "heatmap_generated": False,
            "bounding_boxes": [],
            "message": "Grad-CAM integration pending model deployment."
        }

image_explainer_service = ImageExplainer()
