requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.4
pydantic-settings==2.2.1
SQLAlchemy==2.0.28
chromadb==0.4.24
sentence-transformers==2.5.1
openai==1.14.0
python-multipart==0.0.9
python-dotenv==1.0.1
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0
loguru==0.7.2
Pillow==10.2.0
